+++ tests/variables.global.test.mjs
@@ -20,7 +20,7 @@ test('variables.global', () => {
     expect(is_undefined).not.toBeUndefined();
     expect(is_callable).not.toBeUndefined();
     expect(to_array).not.toBeUndefined();
-    expect(boolval).not.toBeUndefined();
+    expect(to_bool).not.toBeUndefined();
     expect(floatval).not.toBeUndefined();
     expect(doubleval).not.toBeUndefined();
     expect(empty).not.toBeUndefined();